requirements.txt
asgiref==3.8.1
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
Django==5.1.2
djangorestframework==3.15.2
executing==2.1.0
ipython==8.28.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
prompt_toolkit==3.0.48
pure_eval==0.2.3
Pygments==2.18.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.2
wcwidth==0.2.13
